# management/registry/openapi/openapi.yaml
basePath: /
definitions:
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handler.StatusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  handler.TrafficSummaryResponse:
    properties:
      active_count:
        example: 42
        type: integer
      over_limit_count:
        example: 3
        type: integer
      total_used:
        example: 1073741824
        type: integer
    type: object
...
info:
  contact: {}
  description: Hive management plane API for nodes and subscriptions.
  title: Hive Node Registry API
  version: 0.1.0
paths:
...
  /admin/subscriptions/{id}/reset-traffic:
    post:
      description: 将指定订阅的 traffic_used 重置为 0
      operationId: AdminResetSubscriptionTraffic
      parameters:
      - description: 订阅 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AdminSessionCookie: []
      summary: 重置订阅流量
      tags:
      - admin
...
  /admin/traffic/summary:
    get:
      description: 返回活跃订阅的流量使用总量、活跃数和超限数
      operationId: AdminTrafficSummary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TrafficSummaryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - AdminSessionCookie: []
      summary: 获取流量汇总
      tags:
      - admin
...
schemes:
- https
- http
securityDefinitions:
  AdminSessionCookie:
    description: '"HttpOnly session cookie for admin UI"'
    in: cookie
    name: hive_admin_session
    type: apiKey
  BearerAuth:
    description: '"Authorization: Bearer <API_SECRET>"'
    in: header
    name: Authorization
    type: apiKey
  CustomerSessionCookie:
    description: '"HttpOnly session cookie for customer portal"'
    in: cookie
    name: hive_customer_session
    type: apiKey
swagger: "2.0"